Add standalone scramblesuit and obfs4 client support
Set the necessary Tor environment variables, and add additional startup
command line parameters that are needed to set up the proxy correctly.

Also add the required arguments to the SOCKS5 request that are needed
for the obfs4 and Scramblesuit handshake to succeed.

The default will be to start obfs4proxy in obfs4 client mode listening
on port 45578.

Special thanks to Daniel Kaldor, Ivan Sergeyenko, and Marcus Osobase for
the original implementation. This is a cleaned-up and modified version
of their work.