Tb 13409: Update GoLang to 1.13, updated build script to accept flags, and upped version to 2.0.0 (#1)
* Bump the version to 0.0.8-dev, signifying development towards 0.0.8.

* Bug 24793: Send the correct authorization HTTP header for basic auth.

Apparently I didn't test the "connect via HTTP(s)" proxy with
authentication at all when I added that functionality, so it has been
broken for years.

This should fix it now.

* fixup! Bug 24793: Send the correct authorization HTTP header for basic auth.

* (meek_lite) Explicitly set Content-Length to zero when there is no data to send.

https://gitweb.torproject.org/pluggable-transports/meek.git/commit/?h=0.28&id=0ea861efe5873b517fbd08c4bc48027e4db16a95

* port to a Go module

The biggest win is that we now declare what versions of each dependency
we require to build. This way, building a certain version of obfs4 will
always use the same source code, independent of the master branch of
each dependency.

This is necessary for reproducible builds. On top of that, go.sum
contains checksums of all the transitive dependencies and their modules,
so the build system will also recognise when the source code has been
changed.

Updated the build instructions accordingly. We don't drop support for
earlier Go versions, but those won't get the benefit of reproducible
builds unless we start vendoring the dependencies too.

* Change the canonical upstream repo location to gitlab

This commit changes the upstream repo location to:
  https://gitlab.com/yawning/obfs4.git

Additionally all the non-`main` sub-packages now have an import
comment annotation.  As a matter of courtesy, I will continue to
push to both the existing github.com and git.torproject.org repos
for the foreseeable future, though I reserve the right to stop
doing so at any time.

* Annotate use of deprecated `net/http/httputil` package

This is to silence some of the static analysis tools used in
development.  Despite `http.Client` and `http.Transport` being
suggested as an alternative, there is no way to accomplish current
functionality with either suggested replacement.

See: https://github.com/golang/go/issues/8285

* Clean up static analysis warnings

Mostly but not entirely discarding error return values of things that
can not possibly fail despite the API returning errors.

* Update README.md (No functional changes)

* Fix missing field size in obfs4-spec.txt

Thanks to @SudoHenk on github for pointing out the issue long ago.

* Do the release ritual for obfs4proxy 0.0.8.

* Bump the version to 0.0.9-dev

* transports/meeklite: Cleanups, bugfixes and improvements

 * Properly close the response body on HTTP error.
 * Cleanup close signaling.
 * Write() should return faster on closed connections.

* transports/meeklite: uTLS for ClientHello camouflage

There's still some interesting oddities depending on remote server and
what fingerprint is chosen, but I can watch videos online with the
chosen settings and the TBB Azure bridge.

Note: Despite what people are claiming in the Tor Browser bug tracker
it isn't all that hard to use the built in http client with utls.  And
yes, the `transport.go` code does negotiate correctly in a standalone
test case (apart from compatibility related oddities).

* transports/meeklite: Add `utls` argument to configure behavior

Per dcf:
> As for the TODO, my plan was was to expose a "utls" SOCKS arg
> to make it configurable per bridge, and just reuse the utls
> Client Hello ID names:
>	utls=HelloChrome_Auto

This adds support for all currently supported utls ClientHello IDs
with the following caveats/differences:

 * `none` - Disables using utls entirely, forces `crypto/tls`.
 * `HelloGolang` - Alias of `none`, since using utls is pointless.
 * `HelloCustom` - Omitted as pointless.

* transports/meeklite: Apply http.DefaultTransport timeouts etc

* fixup! transports/meeklite: uTLS for ClientHello camouflage

Fix `getDialTLSAddr` to always return a integer port.  Thanks to dcf for
reporting the issue.

* fixup! Bug 24793: Send the correct authorization HTTP header for basic auth.

* transports/meeklite: Use a modified version of utls

Changes:
 * Use a fork of utls with some compatibility improvements.
 * Switch the default ClientHello profile to `HelloFirefox_Auto`.
 * Add the `HelloChrome_71` profile.

The existing `HelloFirefox_Auto` profile that points to
`HelloFirefox_63` also matches the (common) behavior of Firefox 65,
assuming that 3DES ciphersuites are not disabled.

* transports/meeklite: Add a lightweight HPKP implementation

HPKP is effectively dead as far as a standard goes, but the idea has
merit in certain use cases, this being one of them.

As a TLS MITM essentially will strip whatever obfuscation that the
transport may provide, the digests of the SubjectPublicKeyInfo fields
of the Tor Browser Azure meek host are now hardcoded.

The behavior can be disabled by passing `disableHPKP=true` on the bridge
line, for cases where comaptibility is prefered over security.

* transports/meek_lite: Add an expiry date for HPKP entries

Mostly since the built-in pins will likely become invalid once the
certificates I used to generate them start to expire.

* transports/meek_lite: Switch to pinning MS's CA intermediary certs

This should give me more time before I need to update this.

* fixup! transports/meek_lite: Switch to pinning MS's CA intermediary certs

* transports/meeklite: Enable the utls fork's vartime group feature

* transports/meeklite: Bump the tag for the utls fork

Upstream fixed a bug, so use a tag that has the important parts
cherry-picked.

* Do the release ritual for obfs4proxy-0.0.9

* Bump the version to 0.0.10-dev

* transports/meeklite: Tweak the TLS configuration

* transports/meek_lite: More utls related changes

 * Bump the module import to a new tag
 * Bump the rest of the dependencies while I'm here
 * Add some new fingerprints from upstream
 * Disable my fork's AES timing sidechannel defenses

* Do the release ritual for obfs4proxy-0.0.10

* Bump the version to 0.0.11-dev

* Update my e-mail address (no functional changes)

* obfs4: Alter tear down behavior to be less distinctive

The old behavior closed the connection on handshake failure after:
 * The first N bytes (random on a per-server basis).
 * The first M seconds (random on a per-server basis).

Whichever came first.  As Sergey Frolov kindly points out, depending on
which conditions cause termination, the server will send either a FIN or
a RST.  This change will remove the "amount read" based termination
threshold, so that connections that cause failed handshakes will discard
all data received until the teardown time is reached.

Thanks to Sergey Frolov for bringing this issue to my attention.

* transports/meek_lite: Bump the version of the utls fork

* Do the release ritual for obfs4proxy-0.0.11

* Bump the version to 0.0.12-dev

* TB-13409: Added two required imports and made min required GoLang 1.13

* TB-13409: Updated the android_build.sh build script

* TB-13409: Added README to building obfs4 for Android

* TB-13409: Renamed obfs4proxyVersion (TB defined version) "obfs4TBAndroidVer" to better reflect what it is. Updated version from 1.0.1 to 2.0.0 because there have been major breaking changes.

* TB-13409: Removed the generated binaries

* TB-13409: Update /gitignore to ignore obfs4proxy/pkg. Will not .gitignore the out folder as we may choose to actually push those binaries as a backup when we can't build from source.

* TB-13409: CR feedback and added testing steps to README (thanks to Sean)

* TB-13409: Added original obfs4proxyVersion number back so we can always keep track against the yawning releases